# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat API Gateway",
    "description": "\n    A FastAPI backend that provides a clean REST interface for the LangXchange chat service.\n\n    ## Features\n\n    - **Authentication**: Manage authentication with the external chat API\n    - **Sessions**: Create and manage chat sessions with AI agents\n    - **Messages**: Send messages and receive AI responses\n    - **WebSocket**: Real-time chat via WebSocket\n    - **Configuration**: Switch between external/internal modes at runtime\n    - **Convenience Endpoint**: Quick chat endpoint for simple interactions\n\n    ## Getting Started\n\n    1. Use `/auth/login` to authenticate (or let the API auto-authenticate)\n    2. Select an agent with `POST /agents/select?agent_uuid=...`\n    3. Create a session with `POST /sessions`\n    4. Send messages with `POST /sessions/{session_uuid}/messages`\n    5. Or use `POST /chat` for quick stateless interactions\n    6. Or connect via WebSocket at `/ws/chat/{session_uuid}`\n\n    ## Runtime State\n    After selecting an agent and creating a session, all endpoints will\n    automatically use those values \u2014 no need to pass them repeatedly.\n    Check current state with `GET /config/state`.\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticate with the external chat service.\n\nIf credentials are not provided, uses the default configured credentials.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "External API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create Session",
        "description": "Create a new chat session with the AI agent.\n\nagent_uuid is optional \u2014 if not provided, uses the runtime selected agent\n(from POST /agents/select) or falls back to the default in config.",
        "operationId": "create_session_sessions_post",
        "parameters": [
          {
            "name": "agent_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The agent UUID to use for this session",
              "title": "Agent Uuid"
            },
            "description": "The agent UUID to use for this session"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Session creation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{session_uuid}/messages": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send Message",
        "description": "Send a message to an active chat session.\n\nThe message will be processed by the AI agent and a response will be returned.",
        "operationId": "send_message_sessions__session_uuid__messages_post",
        "parameters": [
          {
            "name": "session_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The session UUID",
              "title": "Session Uuid"
            },
            "description": "The session UUID"
          },
          {
            "name": "agent_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Message sending failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/config/state": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get Runtime State",
        "description": "Get the current runtime state (active agent, session, mode).",
        "operationId": "get_runtime_state_config_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "Set Runtime State",
        "description": "Manually override runtime agent UUID and/or session UUID.\n\nUse this to pin a specific agent or session without going through\nthe select agent / create session flow.",
        "operationId": "set_runtime_state_config_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/chat": {
      "post": {
        "tags": [
          "Quick Chat"
        ],
        "summary": "Quick chat endpoint",
        "description": "Convenience endpoint for quick chat interactions.\n\n- Uses runtime agent UUID (from POST /agents/select) or config fallback\n- Uses runtime session UUID (from POST /sessions) or creates a new one\n- Respects current chat mode (external/internal)\n- If session_uuid is passed in body it takes priority over runtime state",
        "operationId": "quick_chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "title": "AuthRequest",
        "description": "Authentication request model."
      },
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "AuthResponse",
        "description": "Authentication response model."
      },
...
      "ConfigurationsUsed": {
        "properties": {
          "vector_config": {
            "type": "boolean",
            "title": "Vector Config",
            "default": false
          },
          "file_config": {
            "type": "boolean",
            "title": "File Config",
            "default": false
          },
          "database_config": {
            "type": "boolean",
            "title": "Database Config",
            "default": false
          },
          "mcp_config": {
            "type": "boolean",
            "title": "Mcp Config",
            "default": false
          }
        },
        "type": "object",
        "title": "ConfigurationsUsed",
        "description": "Configuration usage status."
      },
      "CreateSessionRequest": {
        "properties": {
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "description": "System prompt for the AI agent",
            "default": "You are a helpful AI assistant. Please provide clear, concise, and helpful responses."
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Unique identifier for the user",
            "default": "default-user"
          },
          "user_prompt": {
            "type": "string",
            "title": "User Prompt",
            "description": "Initial user message",
            "default": "Hello, I just started a new chat session."
          },
          "use_fileconfig": {
            "type": "boolean",
            "title": "Use Fileconfig",
            "default": false
          },
          "use_ddbconfig": {
            "type": "boolean",
            "title": "Use Ddbconfig",
            "default": false
          },
          "use_vectorconfig": {
            "type": "boolean",
            "title": "Use Vectorconfig",
            "default": true
          },
          "use_mcpconfig": {
            "type": "boolean",
            "title": "Use Mcpconfig",
            "default": false
          },
          "use_remote_storage": {
            "type": "boolean",
            "title": "Use Remote Storage",
            "default": false
          }
        },
        "type": "object",
        "title": "CreateSessionRequest",
        "description": "Request model for creating a chat session."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Error response model."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMConfig": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          }
        },
        "type": "object",
        "title": "LLMConfig",
        "description": "LLM configuration details."
      },
      "MessageResponse": {
        "properties": {
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "processing_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Ms"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "MessageResponse",
        "description": "Response model for message operations."
      },
      "QuickChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "default-user"
          },
          "session_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Uuid"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "default": "You are a helpful AI assistant."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "QuickChatRequest",
        "description": "Quick chat request model."
      },
      "QuickChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "session_uuid": {
            "type": "string",
            "title": "Session Uuid"
          },
          "processing_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Ms"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          }
        },
        "type": "object",
        "required": [
          "response",
          "session_uuid"
        ],
        "title": "QuickChatResponse",
        "description": "Quick chat response model."
      },
      "RuntimeStateRequest": {
        "properties": {
          "agent_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Uuid"
          },
          "session_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Uuid"
          }
        },
        "type": "object",
        "title": "RuntimeStateRequest"
      },
      "RuntimeStateResponse": {
        "properties": {
          "agent_uuid": {
            "type": "string",
            "title": "Agent Uuid"
          },
          "session_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Uuid"
          },
          "chat_mode": {
            "type": "string",
            "enum": [
              "external",
              "internal"
            ],
            "title": "Chat Mode"
          }
        },
        "type": "object",
        "required": [
          "agent_uuid",
          "session_uuid",
          "chat_mode"
        ],
        "title": "RuntimeStateResponse"
      },
      "SendMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "The message content to send"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SendMessageRequest",
        "description": "Request model for sending a message."
      },
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "session_uuid": {
            "type": "string",
            "title": "Session Uuid"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "agent_uuid": {
            "type": "string",
            "title": "Agent Uuid"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "llm_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "configurations_used": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigurationsUsed"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "session_uuid",
          "agent_name",
          "agent_uuid"
        ],
        "title": "SessionResponse",
        "description": "Response model for session operations."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}